+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [1.3.0] - 2025-08-22
+
 ### Changed
 
 - Codemod now replaces top-level API when they have args, such as `z.string().url("Please enter a valid URL")` => `z.url("Please enter a valid URL")`. ([#18](https://github.com/nicoespeon/zod-v3-to-v4/pull/18))
@@ -172,6 +174,7 @@ For more information about Zod v4 changes, see the [official migration guide](ht
 
 <!-- Links -->
 
+[1.3.0]: https://github.com/nicoespeon/zod-v3-to-v4/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/nicoespeon/zod-v3-to-v4/compare/1.1.5...1.2.0
 [1.1.5]: https://github.com/nicoespeon/zod-v3-to-v4/compare/1.1.4...1.1.5
 [1.1.4]: https://github.com/nicoespeon/zod-v3-to-v4/compare/1.1.3...1.1.4